+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # Changelog
+## 3.0.0 - 16/07/2022
+* Refactor infrastructure
+  * Add IAM policy per resource
+  * Add output variables
+  * Merge tags in local declaration
+  * Add KMS key and alias
+  * Add bucket public access block
+  * Add bucket logging
+  * Add bucket ACL
+  * Add bucket versioning
+  * Add bucket serverside encryption
+  * Add bucket SSL access policy
+
 ## 2.0.0 - 14/03/2021
 * Add server side encryption to bucket objects
 * Create a user to be used when provisioning infrastructure